Updated tests

+++ lib/match/git_helper.rb
@@ -52,7 +52,7 @@ module Match
     end
 
     def self.clear_changes
-      FileUtils.rm_rf(@dir)
+      FileUtils.rm_rf(@dir) if @dir # @dir might be nil in tests
       UI.success "🔒  Successfully encrypted certificates repo" # so the user is happy
       @dir = nil
     end
